src/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-filter==2.2.0
firebase-admin==4.2.0
gunicorn==20.0.4
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
